consolidate developer docs in gen; enter only emits its own schema
Enter's docs.ts shrinks from 1460 to 29 lines. All schema decoration
(info description, tags, alias filtering, media merge, x-codeSamples,
response examples) moves to gen and runs post-merge. Enter now exposes
just two routes: a 301 redirect to gen for /docs/, and the raw schema
endpoint that gen consumes via service binding.

Also normalizes two route tags from "Account" to "👤 Account" so they
group correctly in Scalar.

Fixes #11296
